【NPU】Fix reshape test & add grad test (#31776)
* fix

* fix
